On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   backend/alembic/__pycache__/env.cpython-311.pyc
	new file:   backend/alembic/versions/6beb8cc60967_updated_models_to_include_first_name_.py
	modified:   backend/alembic/versions/__pycache__/5e69cc67b620_made_changes_to_delete_sightings_when_.cpython-311.pyc
	new file:   backend/alembic/versions/__pycache__/6beb8cc60967_updated_models_to_include_first_name_.cpython-311.pyc
	modified:   backend/models/dives.py
	modified:   backend/models/user_preferences.py
	modified:   backend/models/users.py
	modified:   backend/routes/db_routes.py
	modified:   backend/services/user_service.py
	modified:   backend/tests/test_dive_model.py
	modified:   backend/tests/test_user_model.py
	modified:   frontend/src/containers/PreviousEntries.js
	modified:   frontend/src/containers/Register.js
	modified:   frontend/src/utils/convertUnits.js

+++ backend/services/user_service.py
@@ -1,8 +1,9 @@
 from models.users import User
+from models.user_preferences import UserPreferences
 from extensions import db
 
 
-def create_user(username, email, password, is_approved=False, admin=False):
+def create_user(username, email, password, first_name=None, preferred_units='imperial', is_approved=False, admin=False): 
     # Create a new user
     user = User(
         username=username,
@@ -11,7 +12,18 @@ def create_user(username, email, password, is_approved=False, admin=False):
         is_approved=is_approved,
         admin=admin
     )
+    
     db.session.add(user)
+    
+    # Create a new user preferences
+    user_preferences = UserPreferences(
+        user_id=user.id,
+        first_name=first_name,
+        preferred_units=preferred_units
+    )
+    
+    db.session.add(user_preferences)
+    
     db.session.commit()
 
     return user
@@ -34,6 +46,8 @@ def get_user_by_email(email):
 
 def delete_user(user):
     # Delete the user from the database
+    if user is None:
+        return
     db.session.delete(user)
     db.session.commit()
 
@@ -44,19 +58,30 @@ def get_all_users():
 
 def update_username(user, new_username):
     # Update the user's username
+    if user is None:
+        return
     if User.validate_username_availability(new_username):
         user.username = new_username
         db.session.commit()
     
 def update_email(user, new_email):
     # Update the user's email
+    if user is None:
+        return
     if User.validate_email_availability(new_email) and User.validate_email_format(new_email):
         user.email = new_email
         db.session.commit()
     
 def change_password(user, new_password):
     # Update the user's password
+    if user is None:
+        return
     if User.validate_password_strength(new_password):
         user.password = new_password
         db.session.commit()
     
+def get_user_preferences(user):
+    # Retrieve the user's preferences from the database
+    if user is None:
+        return
+    return user.user_preferences.get_user_preferences()
